Fix warning in multiple application tests
After 2296989 the block passed to the application's initialize method is
evaluated when run_load_hooks! is called. Since this only happens when
the singleton instance is called any block passed to a second application
never be evaluated. Therefore we need to set eager_load to false on the
application instance before initializing to fix the warning.